add PtGenSite and parse json with subtypes.

+++ internet/src/main/java/wsg/tools/internet/video/entity/douban/pojo/Rating.java
@@ -15,11 +15,12 @@ import java.util.Map;
 @Setter
 @Getter
 public class Rating {
-    private Integer max;
+
+    private Double max;
     private Double average;
     private Map<Integer, Integer> details;
     @JsonAlias("numRaters")
     private Integer stars;
-    private Integer min;
-    private Integer value;
+    private Double min;
+    private Double value;
 }
